feat(utils): Move envelope utils to `@sentry/utils`
