--- packages/core/src/envelope.ts
@@ -1,88 +1,0 @@
-import type {
-  DsnComponents,
-  Event,
-  EventEnvelope,
-  EventItem,
-  SdkInfo,
-  SdkMetadata,
-  Session,
-  SessionAggregates,
-  SessionEnvelope,
-  SessionItem,
-} from '@sentry/types';
-import {
-  createEnvelope,
-  createEventEnvelopeHeaders,
-  dsnToString,
-  getSdkMetadataForEnvelopeHeader,
-} from '@sentry/utils';
-
-/**
- * Apply SdkInfo (name, version, packages, integrations) to the corresponding event key.
- * Merge with existing data if any.
- **/
-function enhanceEventWithSdkInfo(event: Event, sdkInfo?: SdkInfo): Event {
-  if (!sdkInfo) {
-    return event;
-  }
-  event.sdk = event.sdk || {};
-  event.sdk.name = event.sdk.name || sdkInfo.name;
-  event.sdk.version = event.sdk.version || sdkInfo.version;
-  event.sdk.integrations = [...(event.sdk.integrations || []), ...(sdkInfo.integrations || [])];
-  event.sdk.packages = [...(event.sdk.packages || []), ...(sdkInfo.packages || [])];
-  return event;
-}
-
-/** Creates an envelope from a Session */
-export function createSessionEnvelope(
-  session: Session | SessionAggregates,
-  dsn?: DsnComponents,
-  metadata?: SdkMetadata,
-  tunnel?: string,
-): SessionEnvelope {
-  const sdkInfo = getSdkMetadataForEnvelopeHeader(metadata);
-  const envelopeHeaders = {
-    sent_at: new Date().toISOString(),
-    ...(sdkInfo && { sdk: sdkInfo }),
-    ...(!!tunnel && dsn && { dsn: dsnToString(dsn) }),
-  };
-
-  const envelopeItem: SessionItem =
-    'aggregates' in session ? [{ type: 'sessions' }, session] : [{ type: 'session' }, session.toJSON()];
-
-  return createEnvelope<SessionEnvelope>(envelopeHeaders, [envelopeItem]);
-}
-
-/**
- * Create an Envelope from an event.
- */
-export function createEventEnvelope(
-  event: Event,
-  dsn?: DsnComponents,
-  metadata?: SdkMetadata,
-  tunnel?: string,
-): EventEnvelope {
-  const sdkInfo = getSdkMetadataForEnvelopeHeader(metadata);
-
-  /*
-    Note: Due to TS, event.type may be `replay_event`, theoretically.
-    In practice, we never call `createEventEnvelope` with `replay_event` type,
-    and we'd have to adjut a looot of types to make this work properly.
-    We want to avoid casting this around, as that could lead to bugs (e.g. when we add another type)
-    So the safe choice is to really guard against the replay_event type here.
-  */
-  const eventType = event.type && event.type !== 'replay_event' ? event.type : 'event';
-
-  enhanceEventWithSdkInfo(event, metadata && metadata.sdk);
-
-  const envelopeHeaders = createEventEnvelopeHeaders(event, sdkInfo, tunnel, dsn);
-
-  // Prevent this data (which, if it exists, was used in earlier steps in the processing pipeline) from being sent to
-  // sentry. (Note: Our use of this property comes and goes with whatever we might be debugging, whatever hacks we may
-  // have temporarily added, etc. Even if we don't happen to be using it at some point in the future, let's not get rid
-  // of this `delete`, lest we miss putting it back in the next time the property is in use.)
-  delete event.sdkProcessingMetadata;
-
-  const eventItem: EventItem = [{ type: eventType }, event];
-  return createEnvelope<EventEnvelope>(envelopeHeaders, [eventItem]);
-}
