+++ packages/core/src/index.ts
@@ -5,7 +5,6 @@ export type { ServerRuntimeClientOptions } from './server-runtime-client';
 export type { RequestDataIntegrationOptions } from './integrations/requestdata';
 
 export * from './tracing';
-export { createEventEnvelope } from './envelope';
 export {
   addBreadcrumb,
   captureCheckIn,
